refactor: migrate e2e to plugin
