+++ packages/core/client/src/modules/popup/__e2e__/templatesOfBug.ts
@@ -5124,704 +5124,3 @@ export const zIndexOfSubpage = {
     'x-index': 1,
   },
 };
-export const theAddBlockButtonInDrawerShouldBeVisible = {
-  pageSchema: {
-    _isJSONSchemaObject: true,
-    version: '2.0',
-    type: 'void',
-    'x-component': 'Page',
-    'x-app-version': '1.3.32-beta',
-    properties: {
-      ydtgms2lmr4: {
-        _isJSONSchemaObject: true,
-        version: '2.0',
-        type: 'void',
-        'x-component': 'Grid',
-        'x-initializer': 'page:addBlock',
-        'x-app-version': '1.3.32-beta',
-        properties: {
-          jk8ix7ivmvb: {
-            _isJSONSchemaObject: true,
-            version: '2.0',
-            type: 'void',
-            'x-component': 'Grid.Row',
-            'x-app-version': '1.3.32-beta',
-            properties: {
-              fdohqzh304u: {
-                _isJSONSchemaObject: true,
-                version: '2.0',
-                type: 'void',
-                'x-component': 'Grid.Col',
-                'x-app-version': '1.3.32-beta',
-                properties: {
-                  syp961jrbvs: {
-                    _isJSONSchemaObject: true,
-                    version: '2.0',
-                    type: 'void',
-                    'x-decorator': 'TableBlockProvider',
-                    'x-acl-action': 'users:list',
-                    'x-use-decorator-props': 'useTableBlockDecoratorProps',
-                    'x-decorator-props': {
-                      collection: 'users',
-                      dataSource: 'main',
-                      action: 'list',
-                      params: {
-                        pageSize: 20,
-                      },
-                      rowKey: 'id',
-                      showIndex: true,
-                      dragSort: false,
-                    },
-                    'x-toolbar': 'BlockSchemaToolbar',
-                    'x-settings': 'blockSettings:table',
-                    'x-component': 'CardItem',
-                    'x-filter-targets': [],
-                    'x-app-version': '1.3.32-beta',
-                    properties: {
-                      actions: {
-                        _isJSONSchemaObject: true,
-                        version: '2.0',
-                        type: 'void',
-                        'x-initializer': 'table:configureActions',
-                        'x-component': 'ActionBar',
-                        'x-component-props': {
-                          style: {
-                            marginBottom: 'var(--nb-spacing)',
-                          },
-                        },
-                        'x-app-version': '1.3.32-beta',
-                        'x-uid': 'd9ciqysn98f',
-                        'x-async': false,
-                        'x-index': 1,
-                      },
-                      bflk6vcqbda: {
-                        _isJSONSchemaObject: true,
-                        version: '2.0',
-                        type: 'array',
-                        'x-initializer': 'table:configureColumns',
-                        'x-component': 'TableV2',
-                        'x-use-component-props': 'useTableBlockProps',
-                        'x-component-props': {
-                          rowKey: 'id',
-                          rowSelection: {
-                            type: 'checkbox',
-                          },
-                        },
-                        'x-app-version': '1.3.32-beta',
-                        properties: {
-                          actions: {
-                            _isJSONSchemaObject: true,
-                            version: '2.0',
-                            type: 'void',
-                            title: '{{ t("Actions") }}',
-                            'x-action-column': 'actions',
-                            'x-decorator': 'TableV2.Column.ActionBar',
-                            'x-component': 'TableV2.Column',
-                            'x-toolbar': 'TableColumnSchemaToolbar',
-                            'x-initializer': 'table:configureItemActions',
-                            'x-settings': 'fieldSettings:TableColumn',
-                            'x-toolbar-props': {
-                              initializer: 'table:configureItemActions',
-                            },
-                            'x-app-version': '1.3.32-beta',
-                            properties: {
-                              '9fy397o5n7x': {
-                                _isJSONSchemaObject: true,
-                                version: '2.0',
-                                type: 'void',
-                                'x-decorator': 'DndContext',
-                                'x-component': 'Space',
-                                'x-component-props': {
-                                  split: '|',
-                                },
-                                'x-app-version': '1.3.32-beta',
-                                properties: {
-                                  ujxoicj1do2: {
-                                    'x-uid': 'stie1dytwk9',
-                                    _isJSONSchemaObject: true,
-                                    version: '2.0',
-                                    type: 'void',
-                                    title: 'open subpage',
-                                    'x-action': 'view',
-                                    'x-toolbar': 'ActionSchemaToolbar',
-                                    'x-settings': 'actionSettings:view',
-                                    'x-component': 'Action.Link',
-                                    'x-component-props': {
-                                      openMode: 'page',
-                                      iconColor: '#1677FF',
-                                      danger: false,
-                                    },
-                                    'x-action-context': {
-                                      dataSource: 'main',
-                                      collection: 'users',
-                                    },
-                                    'x-decorator': 'ACLActionProvider',
-                                    'x-designer-props': {
-                                      linkageAction: true,
-                                    },
-                                    properties: {
-                                      drawer: {
-                                        _isJSONSchemaObject: true,
-                                        version: '2.0',
-                                        type: 'void',
-                                        title: '{{ t("View record") }}',
-                                        'x-component': 'Action.Container',
-                                        'x-component-props': {
-                                          className: 'nb-action-popup',
-                                        },
-                                        properties: {
-                                          tabs: {
-                                            _isJSONSchemaObject: true,
-                                            version: '2.0',
-                                            type: 'void',
-                                            'x-component': 'Tabs',
-                                            'x-component-props': {},
-                                            'x-initializer': 'popup:addTab',
-                                            properties: {
-                                              tab1: {
-                                                _isJSONSchemaObject: true,
-                                                version: '2.0',
-                                                type: 'void',
-                                                title: '{{t("Details")}}',
-                                                'x-component': 'Tabs.TabPane',
-                                                'x-designer': 'Tabs.Designer',
-                                                'x-component-props': {},
-                                                properties: {
-                                                  grid: {
-                                                    _isJSONSchemaObject: true,
-                                                    version: '2.0',
-                                                    type: 'void',
-                                                    'x-component': 'Grid',
-                                                    'x-initializer': 'popup:common:addBlock',
-                                                    properties: {
-                                                      gkbkej9bcfg: {
-                                                        _isJSONSchemaObject: true,
-                                                        version: '2.0',
-                                                        type: 'void',
-                                                        'x-component': 'Grid.Row',
-                                                        'x-app-version': '1.3.32-beta',
-                                                        properties: {
-                                                          qsnxw1kyglw: {
-                                                            _isJSONSchemaObject: true,
-                                                            version: '2.0',
-                                                            type: 'void',
-                                                            'x-component': 'Grid.Col',
-                                                            'x-app-version': '1.3.32-beta',
-                                                            properties: {
-                                                              rg5p0jxa04f: {
-                                                                _isJSONSchemaObject: true,
-                                                                version: '2.0',
-                                                                type: 'void',
-                                                                'x-decorator': 'TableBlockProvider',
-                                                                'x-acl-action': 'users:list',
-                                                                'x-use-decorator-props': 'useTableBlockDecoratorProps',
-                                                                'x-decorator-props': {
-                                                                  collection: 'users',
-                                                                  dataSource: 'main',
-                                                                  action: 'list',
-                                                                  params: {
-                                                                    pageSize: 20,
-                                                                  },
-                                                                  rowKey: 'id',
-                                                                  showIndex: true,
-                                                                  dragSort: false,
-                                                                },
-                                                                'x-toolbar': 'BlockSchemaToolbar',
-                                                                'x-settings': 'blockSettings:table',
-                                                                'x-component': 'CardItem',
-                                                                'x-filter-targets': [],
-                                                                'x-app-version': '1.3.32-beta',
-                                                                properties: {
-                                                                  actions: {
-                                                                    _isJSONSchemaObject: true,
-                                                                    version: '2.0',
-                                                                    type: 'void',
-                                                                    'x-initializer': 'table:configureActions',
-                                                                    'x-component': 'ActionBar',
-                                                                    'x-component-props': {
-                                                                      style: {
-                                                                        marginBottom: 'var(--nb-spacing)',
-                                                                      },
-                                                                    },
-                                                                    'x-app-version': '1.3.32-beta',
-                                                                    properties: {
-                                                                      yrh6fyeotth: {
-                                                                        'x-uid': 'm9eykj7hcac',
-                                                                        _isJSONSchemaObject: true,
-                                                                        version: '2.0',
-                                                                        type: 'void',
-                                                                        title: 'Bulk edit 1',
-                                                                        'x-component': 'Action',
-                                                                        'x-action': 'customize:bulkEdit',
-                                                                        'x-action-settings': {
-                                                                          updateMode: 'selected',
-                                                                        },
-                                                                        'x-component-props': {
-                                                                          openMode: 'drawer',
-                                                                          icon: 'EditOutlined',
-                                                                          iconColor: '#1677FF',
-                                                                          danger: false,
-                                                                          type: 'default',
-                                                                        },
-                                                                        'x-align': 'right',
-                                                                        'x-decorator': 'BulkEditActionDecorator',
-                                                                        'x-toolbar': 'ActionSchemaToolbar',
-                                                                        'x-settings': 'actionSettings:bulkEdit',
-                                                                        'x-acl-action': 'update',
-                                                                        'x-acl-action-props': {
-                                                                          skipScopeCheck: true,
-                                                                        },
-                                                                        'x-app-version': '1.3.32-beta',
-                                                                        properties: {
-                                                                          drawer: {
-                                                                            _isJSONSchemaObject: true,
-                                                                            version: '2.0',
-                                                                            type: 'void',
-                                                                            title: '{{t("Bulk edit")}}',
-                                                                            'x-component': 'Action.Container',
-                                                                            'x-component-props': {
-                                                                              className: 'nb-action-popup',
-                                                                            },
-                                                                            'x-app-version': '1.3.32-beta',
-                                                                            properties: {
-                                                                              tabs: {
-                                                                                _isJSONSchemaObject: true,
-                                                                                version: '2.0',
-                                                                                type: 'void',
-                                                                                'x-component': 'Tabs',
-                                                                                'x-component-props': {},
-                                                                                'x-initializer': 'popup:addTab',
-                                                                                'x-initializer-props': {
-                                                                                  gridInitializer:
-                                                                                    'popup:bulkEdit:addBlock',
-                                                                                },
-                                                                                'x-app-version': '1.3.32-beta',
-                                                                                properties: {
-                                                                                  tab1: {
-                                                                                    _isJSONSchemaObject: true,
-                                                                                    version: '2.0',
-                                                                                    type: 'void',
-                                                                                    title: '{{t("Bulk edit")}}',
-                                                                                    'x-component': 'Tabs.TabPane',
-                                                                                    'x-designer': 'Tabs.Designer',
-                                                                                    'x-component-props': {},
-                                                                                    'x-app-version': '1.3.32-beta',
-                                                                                    properties: {
-                                                                                      grid: {
-                                                                                        _isJSONSchemaObject: true,
-                                                                                        version: '2.0',
-                                                                                        type: 'void',
-                                                                                        'x-component': 'Grid',
-                                                                                        'x-initializer':
-                                                                                          'popup:bulkEdit:addBlock',
-                                                                                        'x-app-version': '1.3.32-beta',
-                                                                                        'x-uid': '86kvljipurj',
-                                                                                        'x-async': false,
-                                                                                        'x-index': 1,
-                                                                                      },
-                                                                                    },
-                                                                                    'x-uid': 'osjgnzq0bxu',
-                                                                                    'x-async': false,
-                                                                                    'x-index': 1,
-                                                                                  },
-                                                                                },
-                                                                                'x-uid': 'nddbbs6e5vc',
-                                                                                'x-async': false,
-                                                                                'x-index': 1,
-                                                                              },
-                                                                            },
-                                                                            'x-uid': 'r604iwattq7',
-                                                                            'x-async': false,
-                                                                            'x-index': 1,
-                                                                          },
-                                                                        },
-                                                                        'x-async': false,
-                                                                        'x-index': 1,
-                                                                      },
-                                                                    },
-                                                                    'x-uid': 'kdyyiruax36',
-                                                                    'x-async': false,
-                                                                    'x-index': 1,
-                                                                  },
-                                                                  lodqxwvvrwu: {
-                                                                    _isJSONSchemaObject: true,
-                                                                    version: '2.0',
-                                                                    type: 'array',
-                                                                    'x-initializer': 'table:configureColumns',
-                                                                    'x-component': 'TableV2',
-                                                                    'x-use-component-props': 'useTableBlockProps',
-                                                                    'x-component-props': {
-                                                                      rowKey: 'id',
-                                                                      rowSelection: {
-                                                                        type: 'checkbox',
-                                                                      },
-                                                                    },
-                                                                    'x-app-version': '1.3.32-beta',
-                                                                    properties: {
-                                                                      actions: {
-                                                                        _isJSONSchemaObject: true,
-                                                                        version: '2.0',
-                                                                        type: 'void',
-                                                                        title: '{{ t("Actions") }}',
-                                                                        'x-action-column': 'actions',
-                                                                        'x-decorator': 'TableV2.Column.ActionBar',
-                                                                        'x-component': 'TableV2.Column',
-                                                                        'x-toolbar': 'TableColumnSchemaToolbar',
-                                                                        'x-initializer': 'table:configureItemActions',
-                                                                        'x-settings': 'fieldSettings:TableColumn',
-                                                                        'x-toolbar-props': {
-                                                                          initializer: 'table:configureItemActions',
-                                                                        },
-                                                                        'x-app-version': '1.3.32-beta',
-                                                                        properties: {
-                                                                          '1gt1zh3wlhl': {
-                                                                            _isJSONSchemaObject: true,
-                                                                            version: '2.0',
-                                                                            type: 'void',
-                                                                            'x-decorator': 'DndContext',
-                                                                            'x-component': 'Space',
-                                                                            'x-component-props': {
-                                                                              split: '|',
-                                                                            },
-                                                                            'x-app-version': '1.3.32-beta',
-                                                                            'x-uid': 'r3rllrh7wzk',
-                                                                            'x-async': false,
-                                                                            'x-index': 1,
-                                                                          },
-                                                                        },
-                                                                        'x-uid': 'dmx697o309f',
-                                                                        'x-async': false,
-                                                                        'x-index': 1,
-                                                                      },
-                                                                    },
-                                                                    'x-uid': 'labwg68zvn6',
-                                                                    'x-async': false,
-                                                                    'x-index': 2,
-                                                                  },
-                                                                },
-                                                                'x-uid': '0fn0tjzur2c',
-                                                                'x-async': false,
-                                                                'x-index': 1,
-                                                              },
-                                                            },
-                                                            'x-uid': 'goinlb9fan9',
-                                                            'x-async': false,
-                                                            'x-index': 1,
-                                                          },
-                                                        },
-                                                        'x-uid': '74qg393u7hr',
-                                                        'x-async': false,
-                                                        'x-index': 1,
-                                                      },
-                                                    },
-                                                    'x-uid': 'hetu8763puj',
-                                                    'x-async': false,
-                                                    'x-index': 1,
-                                                  },
-                                                },
-                                                'x-uid': '4dn6kb0cmwp',
-                                                'x-async': false,
-                                                'x-index': 1,
-                                              },
-                                              gljvwj6dtl8: {
-                                                _isJSONSchemaObject: true,
-                                                version: '2.0',
-                                                type: 'void',
-                                                title: 'new tab',
-                                                'x-component': 'Tabs.TabPane',
-                                                'x-designer': 'Tabs.Designer',
-                                                'x-component-props': {},
-                                                'x-app-version': '1.3.32-beta',
-                                                properties: {
-                                                  grid: {
-                                                    _isJSONSchemaObject: true,
-                                                    version: '2.0',
-                                                    type: 'void',
-                                                    'x-component': 'Grid',
-                                                    'x-initializer': 'popup:common:addBlock',
-                                                    'x-app-version': '1.3.32-beta',
-                                                    properties: {
-                                                      c1yhos175hj: {
-                                                        _isJSONSchemaObject: true,
-                                                        version: '2.0',
-                                                        type: 'void',
-                                                        'x-component': 'Grid.Row',
-                                                        'x-app-version': '1.3.32-beta',
-                                                        properties: {
-                                                          yhbbftd4j86: {
-                                                            _isJSONSchemaObject: true,
-                                                            version: '2.0',
-                                                            type: 'void',
-                                                            'x-component': 'Grid.Col',
-                                                            'x-app-version': '1.3.32-beta',
-                                                            properties: {
-                                                              '4y4cdnzx645': {
-                                                                _isJSONSchemaObject: true,
-                                                                version: '2.0',
-                                                                type: 'void',
-                                                                'x-decorator': 'TableBlockProvider',
-                                                                'x-acl-action': 'users:list',
-                                                                'x-use-decorator-props': 'useTableBlockDecoratorProps',
-                                                                'x-decorator-props': {
-                                                                  collection: 'users',
-                                                                  dataSource: 'main',
-                                                                  action: 'list',
-                                                                  params: {
-                                                                    pageSize: 20,
-                                                                  },
-                                                                  rowKey: 'id',
-                                                                  showIndex: true,
-                                                                  dragSort: false,
-                                                                },
-                                                                'x-toolbar': 'BlockSchemaToolbar',
-                                                                'x-settings': 'blockSettings:table',
-                                                                'x-component': 'CardItem',
-                                                                'x-filter-targets': [],
-                                                                'x-app-version': '1.3.32-beta',
-                                                                properties: {
-                                                                  actions: {
-                                                                    _isJSONSchemaObject: true,
-                                                                    version: '2.0',
-                                                                    type: 'void',
-                                                                    'x-initializer': 'table:configureActions',
-                                                                    'x-component': 'ActionBar',
-                                                                    'x-component-props': {
-                                                                      style: {
-                                                                        marginBottom: 'var(--nb-spacing)',
-                                                                      },
-                                                                    },
-                                                                    'x-app-version': '1.3.32-beta',
-                                                                    properties: {
-                                                                      vp75269u5ai: {
-                                                                        'x-uid': '3dup1gvthik',
-                                                                        _isJSONSchemaObject: true,
-                                                                        version: '2.0',
-                                                                        type: 'void',
-                                                                        title: 'Bulk edit 2',
-                                                                        'x-component': 'Action',
-                                                                        'x-action': 'customize:bulkEdit',
-                                                                        'x-action-settings': {
-                                                                          updateMode: 'selected',
-                                                                        },
-                                                                        'x-component-props': {
-                                                                          openMode: 'drawer',
-                                                                          icon: 'EditOutlined',
-                                                                          iconColor: '#1677FF',
-                                                                          danger: false,
-                                                                          type: 'default',
-                                                                        },
-                                                                        'x-align': 'right',
-                                                                        'x-decorator': 'BulkEditActionDecorator',
-                                                                        'x-toolbar': 'ActionSchemaToolbar',
-                                                                        'x-settings': 'actionSettings:bulkEdit',
-                                                                        'x-acl-action': 'update',
-                                                                        'x-acl-action-props': {
-                                                                          skipScopeCheck: true,
-                                                                        },
-                                                                        'x-app-version': '1.3.32-beta',
-                                                                        properties: {
-                                                                          drawer: {
-                                                                            _isJSONSchemaObject: true,
-                                                                            version: '2.0',
-                                                                            type: 'void',
-                                                                            title: '{{t("Bulk edit")}}',
-                                                                            'x-component': 'Action.Container',
-                                                                            'x-component-props': {
-                                                                              className: 'nb-action-popup',
-                                                                            },
-                                                                            'x-app-version': '1.3.32-beta',
-                                                                            properties: {
-                                                                              tabs: {
-                                                                                _isJSONSchemaObject: true,
-                                                                                version: '2.0',
-                                                                                type: 'void',
-                                                                                'x-component': 'Tabs',
-                                                                                'x-component-props': {},
-                                                                                'x-initializer': 'popup:addTab',
-                                                                                'x-initializer-props': {
-                                                                                  gridInitializer:
-                                                                                    'popup:bulkEdit:addBlock',
-                                                                                },
-                                                                                'x-app-version': '1.3.32-beta',
-                                                                                properties: {
-                                                                                  tab1: {
-                                                                                    _isJSONSchemaObject: true,
-                                                                                    version: '2.0',
-                                                                                    type: 'void',
-                                                                                    title: '{{t("Bulk edit")}}',
-                                                                                    'x-component': 'Tabs.TabPane',
-                                                                                    'x-designer': 'Tabs.Designer',
-                                                                                    'x-component-props': {},
-                                                                                    'x-app-version': '1.3.32-beta',
-                                                                                    properties: {
-                                                                                      grid: {
-                                                                                        _isJSONSchemaObject: true,
-                                                                                        version: '2.0',
-                                                                                        type: 'void',
-                                                                                        'x-component': 'Grid',
-                                                                                        'x-initializer':
-                                                                                          'popup:bulkEdit:addBlock',
-                                                                                        'x-app-version': '1.3.32-beta',
-                                                                                        'x-uid': 'ev7zy0yictp',
-                                                                                        'x-async': false,
-                                                                                        'x-index': 1,
-                                                                                      },
-                                                                                    },
-                                                                                    'x-uid': 'n6f6k6fvukf',
-                                                                                    'x-async': false,
-                                                                                    'x-index': 1,
-                                                                                  },
-                                                                                },
-                                                                                'x-uid': 'xm9xscjd3am',
-                                                                                'x-async': false,
-                                                                                'x-index': 1,
-                                                                              },
-                                                                            },
-                                                                            'x-uid': 'tpukizpyz29',
-                                                                            'x-async': false,
-                                                                            'x-index': 1,
-                                                                          },
-                                                                        },
-                                                                        'x-async': false,
-                                                                        'x-index': 1,
-                                                                      },
-                                                                    },
-                                                                    'x-uid': 'mf55d2cpdpx',
-                                                                    'x-async': false,
-                                                                    'x-index': 1,
-                                                                  },
-                                                                  p5i670h31zl: {
-                                                                    _isJSONSchemaObject: true,
-                                                                    version: '2.0',
-                                                                    type: 'array',
-                                                                    'x-initializer': 'table:configureColumns',
-                                                                    'x-component': 'TableV2',
-                                                                    'x-use-component-props': 'useTableBlockProps',
-                                                                    'x-component-props': {
-                                                                      rowKey: 'id',
-                                                                      rowSelection: {
-                                                                        type: 'checkbox',
-                                                                      },
-                                                                    },
-                                                                    'x-app-version': '1.3.32-beta',
-                                                                    properties: {
-                                                                      actions: {
-                                                                        _isJSONSchemaObject: true,
-                                                                        version: '2.0',
-                                                                        type: 'void',
-                                                                        title: '{{ t("Actions") }}',
-                                                                        'x-action-column': 'actions',
-                                                                        'x-decorator': 'TableV2.Column.ActionBar',
-                                                                        'x-component': 'TableV2.Column',
-                                                                        'x-toolbar': 'TableColumnSchemaToolbar',
-                                                                        'x-initializer': 'table:configureItemActions',
-                                                                        'x-settings': 'fieldSettings:TableColumn',
-                                                                        'x-toolbar-props': {
-                                                                          initializer: 'table:configureItemActions',
-                                                                        },
-                                                                        'x-app-version': '1.3.32-beta',
-                                                                        properties: {
-                                                                          owyte0l5cgt: {
-                                                                            _isJSONSchemaObject: true,
-                                                                            version: '2.0',
-                                                                            type: 'void',
-                                                                            'x-decorator': 'DndContext',
-                                                                            'x-component': 'Space',
-                                                                            'x-component-props': {
-                                                                              split: '|',
-                                                                            },
-                                                                            'x-app-version': '1.3.32-beta',
-                                                                            'x-uid': 'bvc3y7y3fem',
-                                                                            'x-async': false,
-                                                                            'x-index': 1,
-                                                                          },
-                                                                        },
-                                                                        'x-uid': '4xso516rjk9',
-                                                                        'x-async': false,
-                                                                        'x-index': 1,
-                                                                      },
-                                                                    },
-                                                                    'x-uid': 'dyyuo828wdo',
-                                                                    'x-async': false,
-                                                                    'x-index': 2,
-                                                                  },
-                                                                },
-                                                                'x-uid': 'um86bfrnd0n',
-                                                                'x-async': false,
-                                                                'x-index': 1,
-                                                              },
-                                                            },
-                                                            'x-uid': 'bwkvo3771d1',
-                                                            'x-async': false,
-                                                            'x-index': 1,
-                                                          },
-                                                        },
-                                                        'x-uid': '8ecnnc4yvgz',
-                                                        'x-async': false,
-                                                        'x-index': 1,
-                                                      },
-                                                    },
-                                                    'x-uid': '6pmwqi2etjn',
-                                                    'x-async': false,
-                                                    'x-index': 1,
-                                                  },
-                                                },
-                                                'x-uid': 'e88jnt5tv4e',
-                                                'x-async': false,
-                                                'x-index': 2,
-                                              },
-                                            },
-                                            'x-uid': '2u9lxm7qivt',
-                                            'x-async': false,
-                                            'x-index': 1,
-                                          },
-                                        },
-                                        'x-uid': 'kqabk7m6d7h',
-                                        'x-async': false,
-                                        'x-index': 1,
-                                      },
-                                    },
-                                    'x-async': false,
-                                    'x-index': 1,
-                                  },
-                                },
-                                'x-uid': 'eqyqzt5letc',
-                                'x-async': false,
-                                'x-index': 1,
-                              },
-                            },
-                            'x-uid': '7t4bewrinmr',
-                            'x-async': false,
-                            'x-index': 1,
-                          },
-                        },
-                        'x-uid': 'ouboqexioby',
-                        'x-async': false,
-                        'x-index': 2,
-                      },
-                    },
-                    'x-uid': 'e363l9dbg2g',
-                    'x-async': false,
-                    'x-index': 1,
-                  },
-                },
-                'x-uid': '9p4tvagqpfd',
-                'x-async': false,
-                'x-index': 1,
-              },
-            },
-            'x-uid': 'a68euip3d3s',
-            'x-async': false,
-            'x-index': 1,
-          },
-        },
-        'x-uid': 'hklfkdijm58',
-        'x-async': false,
-        'x-index': 1,
-      },
-    },
-    'x-uid': '536m19repzv',
-    'x-async': true,
-    'x-index': 1,
-  },
-};
