+++ packages/plugins/@nocobase/plugin-action-bulk-edit/src/client/__e2e__/popup.test.ts
@@ -8,7 +8,7 @@
  */
 
 import { expect, test } from '@nocobase/test/e2e';
-import { theAddBlockButtonInDrawerShouldBeVisible } from './templatesOfBug';
+import { theAddBlockButtonInDrawerShouldBeVisible } from './utils';
 
 test.describe('popup of bulk edit', () => {
   test('the Add block button in drawer should be visible', async ({ page, mockPage }) => {
